Specify the methods in the Tree Node.(In Node.Java) All used math methods will be stored in UsedMath.java as static methods.

+++ src/com/zxx/lyz/Data.java
@@ -4,18 +4,18 @@ package com.zxx.lyz;
  * Created by 李沅泽 on 2016/10/23.
  */
 public class Data {
-    private int[] attr;
+    private double[] attr;
     private int label;
 
     public Data(int dimension) {
-        attr = new int[dimension];
+        attr = new double[dimension];
     }
 
-    public int[] getAttr() {
+    public double[] getAttr() {
         return attr;
     }
 
-    public void setAttr(int[] attr) {
+    public void setAttr(double[] attr) {
         this.attr = attr;
     }
